trans. Prob. Matrix & others

+++ baseVector.c
@@ -6,12 +6,12 @@ int i;
 double dbase; // the base in double type
 double buff;
 
-int baseVector(int length, int base, long int **multipliers){
+int baseVector(int length, int base, int **multipliers){
   dbase = (double) base;
-  *multipliers = (long int *) malloc(length);
+  *multipliers = (int *) malloc(length);
   for(i=0; i<length; i++){
     buff = pow(dbase, ((double) i));
-    multipliers[0][i] = (long int) buff;
+    multipliers[0][i] = (int) buff;
   }
  
   return 0;
