+++ stateId.c
@@ -3,9 +3,9 @@
 #include <math.h>
 
 int i;
-long int state;
+int state;
 
-long int stateId(int *vector, long int *baseVector, int length) {
+int stateId(int *vector, int *baseVector, int length) {
   state = 0;
   
   for(i=0; i<length; i++){
